feat : 매개변수와 관계없이 usecase 실행

+++ RankIn/RankIn/Domain/Model/FetchRankComponents.swift
@@ -10,7 +10,7 @@ import Foundation
 struct FetchRankComponents {
     
     let major: String?
-    let cursorPoint: Int
+    let cursorPoint: Double
     let cursorUserID: String
     
 }
